p29.19 refactor: thread ids.ReconciliationID through every layer
Typed-ID capstone, commit 2: ReconciliationID (the narrowest entity id) is now
a distinct type end-to-end — sqlc model + query funcs, store, reports, web.
Establishes the per-type pattern for the remaining ids.

Enabling fact validated on this first sqlc-model-field typing: database/sql
scans a DB integer into a *ids.ReconciliationID and accepts a typed id as a
query arg through reflection, so typing a model field + query param propagates
with no explicit wraps in the query-func body. Confirmed by building sqlc alone
then running the recon store test.

Threaded:
- sqlc: Reconciliation.ID, ReconciliationsVersion.EntityID (the typed entity
  id); recon-id query params/returns (GetReconciliation, InsertReconciliation
  return, {SetReconciliationStatus,SetReconciliationStatement,HasLaterFinalized-
  Reconciliation,PriorFinalizedStatementBalance}.ID, ReconciliationClearedSum,
  FinalizedReconciliationsForAccountRow.ID, InsertReconciliationVersion ID_2).
- store: every reconID/id param + StartReconciliation return + insertReconcilia-
  tionVersion entityID + FinalizedReconciliation.ID domain field.
- reports: Params.Reconciliation (deferred from p29.18) + a reports.Reconcilia-
  tionID alias; reconciliation_statement.go flows through unchanged.
- web: reconcile.go + reports.go convert at the HTTP-parse seam
  (ids.ReconciliationID(parseID)) and read typed store fields into int64 model/
  path-render values; paramsForm.Recon stays int64 (template render).
- synth/fixture: DemoIDs.OpenRecon, IDs.CheckingUSRecon, ReconciliationExpected.
  ID typed (contained: few consumers).

Boundary discipline (Rule 15): the nullable FK Split.ReconciliationID stays
sql.NullInt64 in the model (scan boundary) and becomes typed only at the store
seam via ids.Null(&reconID) on write and int64(reconID) on compare. NOT touched
(their own later commits): Reconciliation.AccountID, StartReconciliation's
accountID, ReconciliationsVersion.ID (audit-row PK), the Split model. Generic
testutil.AssertVersioned/LatestVersionActor stay int64 (entity-agnostic);
callers pass int64(recon).

make lint (0 issues) + make test (Go + 148 node) + make check green; recon
store/web/reports tests pass; report goldens byte-identical (no testdata churn).

+++ internal/store/reconciliations_test.go
@@ -130,12 +130,12 @@ func TestReconciliationLifecycle(t *testing.T) {
 	if err != nil {
 		t.Fatalf("StartReconciliation: %v", err)
 	}
-	testutil.AssertVersioned(t, e.d, "reconciliations", recon, "create")
+	testutil.AssertVersioned(t, e.d, "reconciliations", int64(recon), "create")
 
 	if err := e.s.SetSplitReconciled(ctx, recon, spID, true); err != nil {
 		t.Fatalf("SetSplitReconciled on: %v", err)
 	}
-	if got := reconOf(t, e.d, spID); got != recon {
+	if got := reconOf(t, e.d, spID); got != int64(recon) {
 		t.Fatalf("split reconciliation_id = %d, want %d", got, recon)
 	}
 	// Clearing is LIVE-ONLY: it mints NO split version (only the create version).
@@ -146,7 +146,7 @@ func TestReconciliationLifecycle(t *testing.T) {
 	if err := e.s.Finalize(ctx, recon); err != nil {
 		t.Fatalf("Finalize: %v", err)
 	}
-	testutil.AssertVersioned(t, e.d, "reconciliations", recon, "update")
+	testutil.AssertVersioned(t, e.d, "reconciliations", int64(recon), "update")
 	got, _ := e.s.GetReconciliation(ctx, recon)
 	if got.Status != "finalized" {
 		t.Errorf("status after Finalize = %q, want finalized", got.Status)
@@ -156,7 +156,7 @@ func TestReconciliationLifecycle(t *testing.T) {
 	if err := e.s.Reopen(ctx, recon); err != nil {
 		t.Fatalf("Reopen: %v", err)
 	}
-	testutil.AssertVersioned(t, e.d, "reconciliations", recon, "update")
+	testutil.AssertVersioned(t, e.d, "reconciliations", int64(recon), "update")
 	got, _ = e.s.GetReconciliation(ctx, recon)
 	if got.Status != "open" {
 		t.Errorf("status after Reopen = %q, want open", got.Status)
@@ -367,7 +367,7 @@ func TestReconciliationReopenAudited(t *testing.T) {
 	if err := e.s.Reopen(reopenCtx, recon); err != nil {
 		t.Fatalf("reopen: %v", err)
 	}
-	if actor := testutil.LatestVersionActor(t, e.d, "reconciliations", recon); actor != reopener {
+	if actor := testutil.LatestVersionActor(t, e.d, "reconciliations", int64(recon)); actor != reopener {
 		t.Errorf("reopen version actor = %d, want %d", actor, reopener)
 	}
 }
@@ -984,7 +984,7 @@ func TestEditReconciliationStatement(t *testing.T) {
 	if err := e.s.EditReconciliationStatement(ctx, recon, "2026-03-02", 25_000); err != nil {
 		t.Fatalf("EditReconciliationStatement: %v", err)
 	}
-	testutil.AssertVersioned(t, e.d, "reconciliations", recon, "update")
+	testutil.AssertVersioned(t, e.d, "reconciliations", int64(recon), "update")
 
 	got, _ := e.s.GetReconciliation(ctx, recon)
 	if got.StatementDate != "2026-03-02" || got.StatementBalance != 25_000 {
@@ -1056,7 +1056,7 @@ func TestDiscardReconciliation(t *testing.T) {
 	if err := e.s.SetSplitReconciled(ctx, recon, spID, true); err != nil {
 		t.Fatalf("clear split: %v", err)
 	}
-	if got := reconOf(t, e.d, spID); got != recon {
+	if got := reconOf(t, e.d, spID); got != int64(recon) {
 		t.Fatalf("split cleared against %d, want %d", got, recon)
 	}
 	preSplitVersions := splitVersionCount(t, e.d, spID)
@@ -1070,7 +1070,7 @@ func TestDiscardReconciliation(t *testing.T) {
 	if err := e.s.DiscardReconciliation(ctx, recon); err != nil {
 		t.Fatalf("DiscardReconciliation: %v", err)
 	}
-	testutil.AssertVersioned(t, e.d, "reconciliations", recon, "update")
+	testutil.AssertVersioned(t, e.d, "reconciliations", int64(recon), "update")
 
 	got, _ := e.s.GetReconciliation(ctx, recon)
 	if got.Status != "discarded" {
